pyproject.toml
[project]
name = "commit-message-benchmark"
version = "0.1.0"
description = "Reproducible benchmark for picking the best LLM to power a prepare-commit-msg hook."
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
bench = "cmb.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cmb"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
